# orchestrateur/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gestion des containers Bridge Archipelago — création, arrêt, suppression et monitoring.",
        "title": "Archilan Orchestrateur API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/apworlds": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stores the apworld in Minio, runs a one-shot Archipelago container to generate\nthe default YAML template, stores it, and returns both the hash and the YAML.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "apworlds"
                ],
                "summary": "Upload an apworld and generate its YAML template",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The .apworld file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_api.UploadApworldResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/apworlds/{hash}/yaml-template": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the default YAML template for the given apworld hash",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "apworlds"
                ],
                "summary": "Get apworld default YAML template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Apworld SHA-256 hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/containers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all managed Bridge containers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "containers"
                ],
                "summary": "List containers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ContainersResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allocates a port, starts a Bridge container, and records it. Returns immediately; webhook fires when container is ready.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "containers"
                ],
                "summary": "Create container",
                "parameters": [
                    {
                        "description": "Session info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateContainerRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_api.CreateContainerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Session already exists",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Port pool exhausted",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/containers/{sessionId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single managed Bridge container by session ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "containers"
                ],
                "summary": "Get container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ContainerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Force-removes a Bridge container and releases its port",
                "tags": [
                    "containers"
                ],
                "summary": "Remove container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/containers/{sessionId}/reload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restarts a Bridge container",
                "tags": [
                    "containers"
                ],
                "summary": "Reload container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/containers/{sessionId}/stop": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends SIGTERM to a running Bridge container",
                "tags": [
                    "containers"
                ],
                "summary": "Stop container",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api.ContainerResponse": {
            "type": "object",
            "properties": {
                "containerId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "sessionId": {
                    "type": "string"
                },
                "status": {
                    "description": "starting | running | stopped | crashed",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "internal_api.ContainersResponse": {
            "type": "object",
            "properties": {
                "containers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.ContainerResponse"
                    }
                }
            }
        },
        "internal_api.CreateContainerRequest": {
            "type": "object",
            "properties": {
                "adminPassword": {
                    "description": "required",
                    "type": "string",
                    "example": "s3cr3t"
                },
                "serverPassword": {
                    "description": "optional — leave empty for open games",
                    "type": "string",
                    "example": ""
                },
                "sessionId": {
                    "type": "string",
                    "example": "weekly-run-2026-05-20"
                }
            }
        },
        "internal_api.CreateContainerResponse": {
            "type": "object",
            "properties": {
                "port": {
                    "type": "integer",
                    "example": 25000
                },
                "sessionId": {
                    "type": "string",
                    "example": "weekly-run-2026-05-20"
                },
                "status": {
                    "type": "string",
                    "example": "starting"
                }
            }
        },
        "internal_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "session not found"
                }
            }
        },
...
        "internal_api.UploadApworldResponse": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string",
                    "example": "0fd8936279e053df96110fcb7898447a9fb8655343b8f26c22108d79a73b4e21"
                },
                "yaml": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003cAPI_KEY\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}